Added kebab-case and changed Names
